In the Parser class: renderPage was extended to check if a requested page, when not found, really was a weblog (but only if using mod_rewrite of course).
In the Weblogs class: Added the function _getLink so the link is always
correctly set (checking for mod_rewrite). This (together with the renderPage
change) fixes bug 507. Introduced the concept of a default weblog (selected
either as the root or the first in the weblog array). Added a the default
weblog check to setCurrentFromCategory to fix a bug where the wrong weblog was
selected.